requirements.txt
# Core ML/LLM Libraries (Match versions from notebook if possible)
torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.1 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu121
transformers==4.41.2
accelerate==0.30.1
peft==0.11.1
trl==0.8.6
bitsandbytes==0.43.1
datasets>=2.1.0 # Use a reasonable recent version
einops

# Data Handling & Plotting
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn # Often useful, linregress is in scipy but sklearn has more

# Utilities
PyYAML>=6.0 # For config loading
tqdm # Progress bars often used by libraries

# Optional: For image display in some environments (like notebooks)
# IPython # Removed as main script likely won't use display()